requirements.txt
ffmpeg-python
openai-whisper
transformers
torch
tqdm
